Contact section (#10)
* feat: add Firebase configuration and setup

* Contact List with Firebase, Styling and Output

Implemented the Contact List to view all contacts. When a firebase
connection is established, the component fetches all information from
the 'contacts' collection and shows a list. Styling with SCSS also done
with width 100% to be implemented in a flex/grid.
When clicked on a item, the component emits a string output, which is
the Id of the collection entry.

* feat: contact-view added

* initials fixed in contact-list component

* test remote and local setup

* Implemented Service for Contact Information

* implemented contact service with fetching for the list and the view

* Add routing to ContactView and make it dynamic based on URL

* Add 2nd approach for the Routing using snapshot

* removed the commented line from app.html

* Added Contacts Component to have Contact List and Contact View inside

---------

Co-authored-by: CalvinH-Dev <hanisch.calvin@gmail.com>
Co-authored-by: Falko Katzer <162439151+hcplayer1988@users.noreply.github.com>